+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+###
+- Pagination in the API
+- Pagination of the results in the webapp
 ### Changed
 - Design improvements:
 	- Change the background color of hovered result rows